+++ src/main/java/us/jonathans/entity/geom/Align.java
@@ -1,4 +1,4 @@
-package us.jonathans.geom;
+package us.jonathans.entity.geom;
 
 public enum Align {
     TOP_LEFT,
