+++ src/main/java/us/jonathans/entity/geom/Vec2.java
@@ -1,4 +1,4 @@
-package us.jonathans.geom;
+package us.jonathans.entity.geom;
 
 public class Vec2 {
     public final int x;
