+++ src/main/java/us/jonathans/entity/sprites/Hole.java
@@ -1,7 +1,7 @@
-package us.jonathans.sprites;
+package us.jonathans.entity.sprites;
 
-import us.jonathans.geom.Align;
-import us.jonathans.geom.CircleBoundingBox;
+import us.jonathans.entity.geom.Align;
+import us.jonathans.entity.geom.CircleBoundingBox;
 
 import java.awt.*;
