+++ src/main/java/us/jonathans/entity/sprites/StoneColors.java
@@ -1,4 +1,4 @@
-package us.jonathans.sprites;
+package us.jonathans.entity.sprites;
 
 import java.awt.*;
 import java.util.Random;
