+++ src/main/java/us/jonathans/use_case/start_game/GameSetterInterface.java
@@ -1,0 +1,4 @@
+package us.jonathans.use_case.start_game;
+
+public class GameSetterInterface {
+}
